Move sentry stuff to its own file
